+++ src/api/index.ts
@@ -4,10 +4,3 @@ export * from './client';
 export * from './errorHandling';
 export * from './operationsList';
 export * from './schema';
-
-/** @deprecated TODO Remove and propagate import changes */
-export * from '../common/fragments/common';
-export * from '../common/approach';
-export * from '../common/displayEnums';
-export * from '../common/secured';
-export * from '../common/sensitivity';
