+++ src/components/BooleanProperty/BooleanProperty.tsx
@@ -2,7 +2,7 @@ import { Chip, ChipProps, makeStyles } from '@material-ui/core';
 import { Skeleton } from '@material-ui/lab';
 import React, { ReactElement } from 'react';
 import { Except, SetRequired } from 'type-fest';
-import { SecuredProp } from '../../api';
+import { SecuredProp } from '~/common';
 import { Redacted } from '../Redacted';
 
 export interface BooleanPropertyProps extends SetRequired<ChipProps, 'label'> {
