+++ src/components/DefinedFileCard/DefinedFileCard.tsx
@@ -15,7 +15,8 @@ import { Skeleton } from '@material-ui/lab';
 import { DateTime } from 'luxon';
 import React, { forwardRef, ReactNode } from 'react';
 import { useDropzone } from 'react-dropzone';
-import { CreateDefinedFileVersionInput, SecuredProp } from '../../api';
+import { CreateDefinedFileVersionInput } from '~/api';
+import { SecuredProp } from '~/common';
 import {
   FileActionsPopup as ActionsMenu,
   FileAction,
