+++ src/components/PeriodicReports/ReportLabel.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import { isSecured } from '~/api';
-import { CalendarDate, Nullable } from '~/common';
+import { CalendarDate, isSecured, Nullable } from '~/common';
 import { Redacted } from '../Redacted';
 import {
   PeriodicReportFragment,
