+++ src/scenes/Engagement/CeremonyCard/CeremonyCard.tsx
@@ -12,7 +12,8 @@ import {
 import { Skeleton } from '@material-ui/lab';
 import clsx from 'clsx';
 import React, { useMemo } from 'react';
-import { canEditAny, UpdateCeremonyInput } from '../../../api';
+import { UpdateCeremonyInput } from '~/api';
+import { canEditAny } from '~/common';
 import { useDialog } from '../../../components/Dialog';
 import { DialogForm } from '../../../components/Dialog/DialogForm';
 import { DateField, FieldGroup, SubmitError } from '../../../components/form';
