+++ src/scenes/Languages/LanguageForm/LanguageForm.tsx
@@ -2,13 +2,8 @@ import { Grid, makeStyles, Typography } from '@material-ui/core';
 import { setIn } from 'final-form';
 import React, { useMemo } from 'react';
 import { Except } from 'type-fest';
-import {
-  canReadAny,
-  CreateLanguage,
-  SensitivityList,
-  UpdateLanguage,
-} from '~/api';
-import { Nullable } from '~/common';
+import { CreateLanguage, SensitivityList, UpdateLanguage } from '~/api';
+import { canReadAny, Nullable } from '~/common';
 import {
   DialogForm,
   DialogFormProps,
