+++ src/scenes/Products/List/ProductList.tsx
@@ -1,6 +1,7 @@
 import { Card, makeStyles, Typography } from '@material-ui/core';
 import React from 'react';
-import { getChangeset, IdFragment } from '../../../api';
+import { getChangeset } from '~/api';
+import { IdFragment } from '~/common';
 import { List, useListQuery } from '../../../components/List';
 import { ProductCard } from '../../../components/ProductCard';
 import { ProductListDocument } from './ProductList.graphql';
