+++ src/scenes/Products/ProductForm/GoalsSection.tsx
@@ -1,7 +1,7 @@
 import { Typography } from '@material-ui/core';
 import { ToggleButton } from '@material-ui/lab';
 import React from 'react';
-import { displayProductTypes } from '../../../api';
+import { displayProductTypes } from '~/common';
 import { EnumField } from '../../../components/form';
 import { productTypes } from './constants';
 import { DefaultAccordion } from './DefaultAccordion';
